Revert to 91b4861f44879326e32d40c113795f64ec6fa7a5
merge-stream doesn't respect the order of the streams provided to it.
this makes it more appropriate for use as a source of parallel streams,
but it's unclear if returning an array to composeStream means the
streams should be ordered (as with ordered-merge-stream) or not.

Rather than handle this with additional options or detection of the
stream type, composeStream will always expect either a single stream or
a falsy value, and have no opinion about the nature or composure of the
stream provided. This allows composeStream to do only one thing.